Deck Creation, api::NotFound, and more macros
The vaultd API server is getting better.  We now have a `done!(r)`
macro, which introspects the enum type of `r` to figure out what HTTP
code to send to the caller.  This gives us JSON all the time, and preps
us for returning errors-as-Objects, and letting Err() be reserved for
truly unexpected, server-side errors.