Initial extract of cardboard.* JS into module
This commit extracts as much of the functionality from the htdocs/
bits of Vault of Cardboard into a new JS module, called `cardboard`.
This allows us to run tests via mocha/chai, and validate behavior
without a browser, and without manual intervention.

Quite a few bugs were fixed in this refactor, and large portions of the
existing Vue.js implementation in htdocs/ was left as-is, or with few
modifications.  Success!