Fix Docker Builds (diesel vs. uuid and --locked)
A lot has happened since the last jaunt through Docker-land.

For starters, the `vaultd` and `rescry` binaries are gone, and are
instead replaced with dispatch through a single multi-call binary
named `cardboard`.  This simplifies the Docker footprint, since we only
have one image to build (the one wrapping `cardboard` as an entrypoint).

Also, we have to use the `--locked` flag to get `cargo install` to honor
the Cargo.lock file it finds locally. It's a whole big thing. Google it.

In the process of chasing down this issue, I went ahead and updated
Diesel to 1.4.x, and brought UUID 0.8.x (for, ironically, uuidv07) along
for the ride.