Handle Collector Number in CDIF
We can now specify collector numbers in CDIF imports, which means I can
finally be more explicit about which of the 4 variants of the same card
I got from WotC.

This is fully integrated into the Collection Clarifier, in that it
attempts to selectively determine if there are multiple cards in the
same set named the same thing, and only then include the collector
number in the clarified input.

The new format for CDIF looks like this:

    4x BFZ *260 Swamp

Which indicates not only that we have a Swamp from Battle for Zendikar,
but that we also have **specifically** card #260 (which is a Swamp from
Battle for Zendikar).

Fixes #4