Fix Broken Split Clarification
If we ever clarified to more than a single replacement, the transaction
form would fail pretty hardcore, treating an array as a string and
automatically (and poorly!) stringifying it as a comma-delimited string.

Now, we use that fancy `...` syntax to expand out the array in-line, and
splice in as a single flat list of lines.

Fixes #8